Output- The sum of the given number is:=8
a and b are the variable,it is like a container or box which contain the integer.
We can change the values of a and b. To get a desire result.
1-what is int?
Int(integer) is a data type which store the integer value it can be positive or the negative.
2-what is return 0?
The function while return 0 is basically returning False as the final value of the function.